Included authentication to favorites routes

+++ server/controller/favoritesCon.js
@@ -4,9 +4,10 @@ export const createFavorite = async (req, res) => {
     const { user_id, bus_id, bus_name, bus_image, bus_rating, bus_price, bus_address, bus_phone } = req.body;
 
     try {
+        const userId = req.user.userId
         const result = await pool.query(
             'INSERT INTO favorites (user_id, bus_id, bus_name, bus_image, bus_rating, bus_price, bus_address, bus_phone) VALUES ($1, $2, $3, $4, $5, $6, $7, $8) RETURNING *',
-            [user_id, bus_id, bus_name, bus_image, bus_rating, bus_price, bus_address, bus_phone]
+            [userId, bus_id, bus_name, bus_image, bus_rating, bus_price, bus_address, bus_phone]
         );
 
         res.json(result.rows[0]);
@@ -20,7 +21,8 @@ export const getFavoriteById = async (req, res) => {
     const fav_id = req.params.id;
 
     try {
-        const result = await pool.query('SELECT * FROM favorites WHERE fav_id = $1', [fav_id]);
+        const userId = req.user.userId
+        const result = await pool.query('SELECT * FROM favorites WHERE fav_id = $1 and user_id = $2', [fav_id, userId]);
 
         if (result.rows.length === 0) {
             res.status(404).json({ error: 'Favorite not found'});
@@ -35,8 +37,8 @@ export const getFavoriteById = async (req, res) => {
 
 export const getAllFavorites = async (req, res) => {
     try {
-      const result = await pool.query('SELECT * FROM favorites');
-
+      const userId = req.user.userId
+      const result = await pool.query('SELECT * FROM favorites WHERE user_id = $1', [userId]);
       res.json(result.rows);
     } catch (error) {
       console.error(error);
@@ -48,7 +50,8 @@ export const deleteFavorite = async (req, res) => {
   const fav_id = req.params.id;
 
   try {
-    const result = await pool.query('DELETE FROM favorites WHERE fav_id = $1 RETURNING *', [fav_id]);
+    const userId = req.user.userId
+    const result = await pool.query('DELETE FROM favorites WHERE fav_id = $1 and user_id = $2 RETURNING *', [fav_id, userId]);
 
     if (result.rows.length === 0) {
       res.status(404).json({ error: 'Favorite not found' });
